🎨 Remove duplicated code + remove the inplace dataframe modification
